refactor(exec): slim session locator
