+++ docs/Tasks/003-refactor-file-size/tasks.md
@@ -18,7 +18,7 @@ description: "Refactor oversized source files: max 500 LOC, aim <300, avoid part
 - [ ] T010 Refactor `CodexAppServerClient` (split into composed internal modules; keep static parsing helpers callable via `CodexAppServerClient.*`).
 - [ ] T011 Refactor `JsonlEventParser` (split into internal parsing components; keep existing behavior and tests).
 - [ ] T012 Refactor `CodexClient` to remove `partial` (replace `[GeneratedRegex]` usage; compose internal helpers).
-- [ ] T013 Refactor `CodexSessionLocator` to remove `partial` (replace `[GeneratedRegex]`; extract helpers).
+- [x] T013 Refactor `CodexSessionLocator` to remove `partial` (replace `[GeneratedRegex]`; extract helpers).
 - [x] T014 Refactor `CodexStructuredOutputExtensions` (split into internal helpers; keep public extensions).
 - [x] T015 Refactor `ResilientCodexAppServerClient` (extract adapter types so the file is <500).
 - [x] T016 Refactor `CodexProcessLauncher` (extract IO/diagnostics helpers; keep tested internals).
